configured gtest for unit testing

+++ src/features/color.cpp
@@ -1,4 +1,4 @@
-#include "../include/color.h"
+#include "../include/color.hpp"
 
 color add(const color& a, const color& b) {
     return color(a.red + b.red, a.green + b.green, a.blue + b.blue);
@@ -8,7 +8,7 @@ color sub(const color& a, const color& b) {
     return color(a.red - b.red, a.green - b.green, a.blue - b.blue);
 }
 
-color scalar_mult(const color& a, const float& num) {
+color scalar_mult(const color& a, const double& num) {
     return color(a.red * num, a.green * num, a.blue * num);
 }
