+++ src/include/tuple.hpp
@@ -3,12 +3,12 @@
 
 // tuple initiates to (0,0,0,0)
 struct tuple {
-    float x;
-    float y;
-    float z;
-    float w;
+    double x;
+    double y;
+    double z;
+    double w;
 
-    tuple(float x_val = 0, float y_val = 0, float z_val = 0, float w_val = 0)
+    tuple(double x_val = 0, double y_val = 0, double z_val = 0, double w_val = 0)
         : x(x_val), y(y_val), z(z_val), w(w_val) {}
 };
 
@@ -16,14 +16,14 @@ struct tuple {
 class point {
 public:
     tuple point_tuple;
-    point(float x, float y, float z) : point_tuple(x, y, z, 1.0) {}
+    point(double x, double y, double z) : point_tuple(x, y, z, 1.0) {}
 };
 
 //for the vector object
 class vector {
 public:
     tuple vector_tuple;
-    vector(float x, float y, float z) : vector_tuple(x, y, z, 0.0) {}
+    vector(double x, double y, double z) : vector_tuple(x, y, z, 0.0) {}
 };
 
 class projectile {
@@ -41,16 +41,16 @@ public:
 };
 
 
-//compare two floats
-bool equal(const float& a, const float& b);
+//compare two doubles
+bool equal(const double& a, const double& b);
 tuple add(const tuple& a, const tuple& b);
 tuple sub(const tuple& a, const tuple& b);
 tuple negate(const tuple& a);
-tuple scalar_mult(const tuple& a, const float& num);
-tuple scalar_div(const tuple& a, const float& num);
-float magnitude(const tuple& a);
+tuple scalar_mult(const tuple& a, const double& num);
+tuple scalar_div(const tuple& a, const double& num);
+double magnitude(const tuple& a);
 tuple normalize(const tuple& a);
-float dot(const tuple& a, const tuple& b);
+double dot(const tuple& a, const tuple& b);
 tuple cross(const tuple& a, const tuple& b);
 
 bool is_vector(const tuple& a);
